+++ src/main/java/org/broadinstitute/hellbender/utils/read/markduplicates/sparkrecords/PairedEnds.java
@@ -10,7 +10,7 @@ public abstract class PairedEnds extends MarkDuplicatesSparkRecord {
     }
 
     public abstract int getFirstStartPosition();
-    public abstract int getScore();
+    public abstract short getScore();
     public abstract boolean isRead1ReverseStrand();
 
     /**
